제동 - 커뮤니티 기능
- 검색 기능 준비 (미완)

- detail.html
1. 게시글 삭제에 alert 추가 (완)
*스크랩, 삭제 버튼의 존재여부에 따라 addEvent 추가
2. 페이지 이전, 이후 버튼 수정필요함(미완)

- 게시글 삭제 시 먼저 댓글, 스크랩 삭제
1. D_BoardService에서 댓글, 스크랩 삭제 (완)
* 여러 Dao를 쓰는데 그 해당 service에는 존재하지 않음.. 이거 물어봐야함 (미완)
2. @transactional 추가 (완)

- BoardController
1. checkOwner 추가

-------------------------------------------------------------------
- updateForm html
1. 헤더 푸터 추가 (미완)

- 상세조회
*눌렀을 때 비회원은 로그인 유도 해야함 (미완)
3. 댓글 페이징 active 복잡해서 우선 보류 (미완)

- 챌린지 게시판 수정
1. 크기가 너무 큼.. 수정할 수 있을지 모르겠지만 우선 체크(미완)

- 커뮤니티 메인 추가
1. 주간 인기글 할 수 있을지 고민

- 페이징 추가
1. 게시글 수정 삭제 시 현재 페이지로 reflesh 추가 (미완)

+++ app/src/main/java/com/bitcamp/testproject/web/controller/BoardController.java
@@ -19,7 +19,6 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.servlet.ModelAndView;
 import com.bitcamp.testproject.service.BoardCommentService;
-import com.bitcamp.testproject.service.BoardReportService;
 import com.bitcamp.testproject.service.BoardService;
 import com.bitcamp.testproject.vo.Board;
 import com.bitcamp.testproject.vo.BoardCategory;
@@ -36,8 +35,6 @@ public class BoardController {
 
   @Autowired
   BoardService boardService;
-  @Autowired
-  BoardReportService boardReportService;
   @Autowired
   BoardCommentService boardCommentService;
 
@@ -70,7 +67,6 @@ public class BoardController {
   }
 
   private String saveAttachedFile(Part file) throws IOException, ServletException {
-    //    List<AttachedFile> attachedFiles = new ArrayList<>();
     String dirPath = sc.getRealPath("/board/files");
     // 첨부파일이 있다면 실행
     if (file.getSize() != 0) {
@@ -80,38 +76,6 @@ public class BoardController {
     }
     return null;
   }
-
-  //  private String saveThumbnailFile(Part file) throws Exception {
-  //    if (file.getSize() == 0) {
-  //      return null;
-  //    }
-  //
-  //    String dirPath = sc.getRealPath("/board/files");
-  //
-  //    String filename = UUID.randomUUID().toString();
-  //    file.write(dirPath + "/" + filename);
-  //
-  //    return filename;
-  //  }
-
-  //  
-  //  private List<AttachedFile> saveAttachedFiles(Part[] files)
-  //      throws IOException, ServletException {
-  //    List<AttachedFile> attachedFiles = new ArrayList<>();
-  //    String dirPath = sc.getRealPath("/board/files");
-  //
-  //    for (Part part : files) {
-  //      if (part.getSize() == 0) {
-  //        continue;
-  //      }
-  //
-  //      String filename = UUID.randomUUID().toString();
-  //      part.write(dirPath + "/" + filename);
-  //      attachedFiles.add(new AttachedFile(filename));
-  //    }
-  //    return attachedFiles;
-  //  }
-  //  
 
   @GetMapping("form")
   public Model form(int cateno, Model model) throws Exception {
@@ -188,7 +152,6 @@ public class BoardController {
     pageMaker.setCri(cri);
     pageMaker.setTotalCount(boardService.countBoardListTotal(no));
 
-
     // 게시글 카테고리 번호 Board객체에 담아서 map객체에 담을 준비 
     Board catenoInBoard = new Board();
     catenoInBoard.setCateno(no);
@@ -202,21 +165,20 @@ public class BoardController {
   }
 
   @GetMapping("delete")
-  public String delete(
-      int no, 
-      HttpSession session) 
-          throws Exception {
+  public String delete(int no, HttpSession session) throws Exception {
 
-    // 돌아갈 게시판카테고리 넘버 찾기
+    // 삭제 후 돌아갈 게시판카테고리 넘버 찾기
     int cateno = boardService.get(no).getCateno();
+    // 회원 유효성 체크(삭제할 회원이 같이 회원인지)
+    checkOwner(no, session);
 
-    //    checkOwner(no, session);
     if (!boardService.delete(no)) {
       throw new Exception("게시글을 삭제할 수 없습니다.");
     }
 
     return "redirect:list?no=" + cateno;
   }
+
 
   @GetMapping("updateForm")
   public Model updateForm(int no, Model model) throws Exception {
@@ -249,12 +211,7 @@ public class BoardController {
   }
 
   @GetMapping("fileDelete")
-  public String boardFileDelete(
-      int no,
-      HttpSession session) 
-          throws Exception {
-
-    //    AttachedFile attachedFile = boardService.getAttachedFile(no); 
+  public String boardFileDelete(int no, HttpSession session) throws Exception {
 
     Member loginMember = (Member) session.getAttribute("loginMember");
     Board board = boardService.get(no);
@@ -268,12 +225,15 @@ public class BoardController {
       throw new Exception("게시글 첨부파일을 삭제할 수 없습니다.");
     }
 
-
     return "redirect:updateForm?no=" + board.getNo();
   }
 
-
-  /////// 제동 메서드 끝 
+  private void checkOwner(int boardNo, HttpSession session) throws Exception {
+    Member loginMember = (Member) session.getAttribute("loginMember");
+    if (boardService.get(boardNo).getWriter().getNo() != loginMember.getNo()) {
+      throw new Exception("게시글 작성자가 아닙니다.");
+    }
+  }
